+++ src/Symfony/Component/Security/Core/User/UserChecker.php
@@ -28,12 +28,8 @@ class UserChecker implements UserCheckerInterface
      */
     public function checkPreAuth(UserInterface $user)
     {
-        if (!$user instanceof AdvancedUserInterface && !$user instanceof User) {
+        if (!$user instanceof User) {
             return;
-        }
-
-        if ($user instanceof AdvancedUserInterface && !$user instanceof User) {
-            @trigger_error(sprintf('Calling "%s()" with an AdvancedUserInterface is deprecated since Symfony 4.1. Create a custom user checker if you wish to keep this functionality.', __METHOD__), E_USER_DEPRECATED);
         }
 
         if (!$user->isAccountNonLocked()) {
@@ -60,12 +56,8 @@ class UserChecker implements UserCheckerInterface
      */
     public function checkPostAuth(UserInterface $user)
     {
-        if (!$user instanceof AdvancedUserInterface && !$user instanceof User) {
+        if (!$user instanceof User) {
             return;
-        }
-
-        if ($user instanceof AdvancedUserInterface && !$user instanceof User) {
-            @trigger_error(sprintf('Calling "%s()" with an AdvancedUserInterface is deprecated since Symfony 4.1. Create a custom user checker if you wish to keep this functionality.', __METHOD__), E_USER_DEPRECATED);
         }
 
         if (!$user->isCredentialsNonExpired()) {